refactor state architecture
