+++ core/src/com/platformer/game/states/gamestates/Multiplayer.java
@@ -1,0 +1,31 @@
+package com.platformer.game.states.gamestates;
+
+import com.badlogic.gdx.graphics.g2d.SpriteBatch;
+import com.platformer.game.states.GameState;
+import com.platformer.game.states.GameStateManager;
+
+public class Multiplayer extends GameState {
+    protected Multiplayer(GameStateManager gsm) {
+        super(gsm);
+    }
+
+    @Override
+    protected void handleInput() {
+        super.handleInput();
+    }
+
+    @Override
+    public void update(float dt) {
+        super.update(dt);
+    }
+
+    @Override
+    public void render(SpriteBatch sb) {
+        super.render(sb);
+    }
+
+    @Override
+    public void dispose() {
+        super.dispose();
+    }
+}
