+++ core/src/com/platformer/game/states/menustates/SettingsMenu.java
@@ -1,0 +1,31 @@
+package com.platformer.game.states.menustates;
+
+import com.badlogic.gdx.graphics.g2d.SpriteBatch;
+import com.platformer.game.states.GameStateManager;
+import com.platformer.game.states.MenuState;
+
+public class SettingsMenu extends MenuState {
+    public SettingsMenu(GameStateManager gsm) {
+        super(gsm);
+    }
+
+    @Override
+    public void handleInput() {
+        super.handleInput();
+    }
+
+    @Override
+    public void update(float dt) {
+        super.update(dt);
+    }
+
+    @Override
+    public void render(SpriteBatch sb) {
+        super.render(sb);
+    }
+
+    @Override
+    public void dispose() {
+        super.dispose();
+    }
+}
